add tf_data node to use tf.data.Dataset instead of feed_dict
passing data to GPU via feed_dict is relatively slow and now
prefetching is possible.
This node creates a tf.data.Dataset and transfers all data
requested from upstream to GPU.

Hint: only request data that is actually needed, otherwise
unnecessary data transfer to GPU performed.

Note: Cannot handle graphs/points. Graph nodes should be
consumed and dropped before reaching this node.
A graph should not be in the global request but should be
requested from the node that consumes it.
This also means that if you want to store them in snapshots,
the snapshot node has to come before this node.